requirements.txt
google-cloud-firestore
google-cloud-datastore
google-cloud-tasks
google-cloud-storage
google-cloud-bigquery